프로젝트 디렉토리 세팅
