+++ src/main/java/com/ohgiraffers/adregamdi/configuration/ContextConfiguration.java
@@ -4,6 +4,6 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ComponentScan("com.ohgiraffers.adregamdi")
+@ComponentScan(basePackages = "com.ohgiraffers.adregamdi")
 public class ContextConfiguration {
 }
