+++ src/main/java/com/ohgiraffers/adregamdi/user/command/domain/aggregate/entity/User.java
@@ -1,0 +1,4 @@
+package com.ohgiraffers.adregamdi.user.command.domain.aggregate.entity;
+
+public class User {
+}
