+++ src/main/java/com/ohgiraffers/adregamdi/user/command/domain/exception/UserException.java
@@ -1,0 +1,4 @@
+package com.ohgiraffers.adregamdi.user.command.domain.exception;
+
+public class UserException {
+}
