+++ src/main/java/com/ohgiraffers/adregamdi/user/command/domain/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.ohgiraffers.adregamdi.user.command.domain.repository;
+
+public interface UserRepository {
+    void insertUser();
+
+    void selectUserOne();
+
+    void selectUserList();
+
+    void updateUser();
+
+    void deleteUser();
+}
