--- src/main/java/com/ohgiraffers/adregamdi/user/domain/entity/User.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.user.domain.entity;
-
-public class User {
-}
