--- src/main/java/com/ohgiraffers/adregamdi/user/domain/exception/UserException.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.user.domain.exception;
-
-public class UserException {
-}
