--- src/main/java/com/ohgiraffers/adregamdi/user/domain/repository/UserRepository.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.user.domain.repository;
-
-public interface UserRepository {
-}
