--- src/main/java/com/ohgiraffers/adregamdi/user/infra/repository/UserInfraRepository.java
@@ -1,8 +1,0 @@
-package com.ohgiraffers.adregamdi.user.infra.repository;
-
-import com.ohgiraffers.adregamdi.user.domain.repository.UserRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class UserInfraRepository implements UserRepository {
-}
