--- src/main/java/com/ohgiraffers/adregamdi/user/infra/service/UserInfraService.java
@@ -1,4 +1,0 @@
-package com.ohgiraffers.adregamdi.user.infra.service;
-
-public class UserInfraService {
-}
