+++ src/main/java/com/ohgiraffers/adregamdi/user/query/infrastructure/repository/UserInfraRepository.java
@@ -1,0 +1,8 @@
+package com.ohgiraffers.adregamdi.user.query.infrastructure.repository;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class UserInfraRepository {
+
+}
