+++ src/main/java/com/ohgiraffers/adregamdi/user/query/infrastructure/service/UserInfraService.java
@@ -1,0 +1,4 @@
+package com.ohgiraffers.adregamdi.user.query.infrastructure.service;
+
+public class UserInfraService {
+}
